+++ conftest.py
@@ -8,9 +8,11 @@ from _pytest.fixtures import FixtureRequest
 from _pytest.tmpdir import TempPathFactory, _mk_tmp
 
 from TestSuite.integration import Integration
+from TestSuite.json_based import JSONBased
 from TestSuite.pack import Pack
 from TestSuite.playbook import Playbook
 from TestSuite.repo import Repo
+from TestSuite.yml import YAML
 
 # Helper Functions
 
@@ -79,6 +81,25 @@ def playbook(request: FixtureRequest, tmp_path_factory: TempPathFactory) -> Play
     """Mocking tmp_path
     """
     return get_playbook(request, tmp_path_factory)
+
+
+@pytest.fixture()
+def malformed_integration_yml(integration) -> YAML:
+    """
+    Provides an invalid integration yml structure.
+    """
+    integration.yml.write("1: 2\n//")
+    return integration.yml
+
+
+@pytest.fixture()
+def malformed_incident_field(pack) -> JSONBased:
+    """
+    Provides an invalid incident field json structure.
+    """
+    incident_field = pack.create_incident_field("malformed")
+    incident_field.write_as_text("{\n '1': '1'")
+    return incident_field
 
 
 @pytest.fixture(scope='session', autouse=True)
