+++ demisto_sdk/commands/doc_reviewer/tests/doc_reviewer_test.py
@@ -1,12 +1,10 @@
 from typing import List
 
-from demisto_sdk.commands.common.tests.tools_test import (  # noqa F401
-    malformed_incident_field, malformed_integration_yml)
 from demisto_sdk.commands.doc_reviewer.doc_reviewer import DocReviewer
 from TestSuite.json_based import JSONBased
 
 
-def test_doc_review_with_release_notes_is_skipped_on_invalid_yml_file(malformed_integration_yml):  # noqa F811
+def test_doc_review_with_release_notes_is_skipped_on_invalid_yml_file(malformed_integration_yml):
     """
     Given -
         malformed yml integration file.
@@ -15,7 +13,7 @@ def test_doc_review_with_release_notes_is_skipped_on_invalid_yml_file(malformed_
         Calling doc-review with --release-notes.
 
     Then -
-        Ensure that no exception/error is raised.
+        Ensure that no exception/error is raised and that the malformed files were not added to the files for review.
     """
     path = malformed_integration_yml.path
 
@@ -27,16 +25,16 @@ def test_doc_review_with_release_notes_is_skipped_on_invalid_yml_file(malformed_
         assert False, str(err)
 
 
-def test_doc_review_with_release_notes_is_skipped_on_invalid_json_file(malformed_incident_field: JSONBased):  # noqa F811
+def test_doc_review_with_release_notes_is_skipped_on_invalid_json_file(malformed_incident_field: JSONBased):
     """
     Given -
-        malformed json incident type.
+        malformed json incident field.
 
     When -
         Calling doc-review with --release-notes.
 
     Then -
-        Ensure that no exception/error is raised.
+        Ensure that no exception/error is raised and that the malformed files were not added to the files for review.
     """
     path = malformed_incident_field.path
 
@@ -48,7 +46,7 @@ def test_doc_review_with_release_notes_is_skipped_on_invalid_json_file(malformed
         assert False, str(err)
 
 
-def test_get_files_from_git_with_invalid_files(mocker, malformed_integration_yml, malformed_incident_field):  # noqa F811
+def test_get_files_from_git_with_invalid_files(mocker, malformed_integration_yml, malformed_incident_field):
     """
     Given -
         malformed json/yml.
